+++ src/main/kotlin/com/github/bgomar/consolelogger/UpdateNotify.kt
@@ -1,14 +1,12 @@
 package com.github.bgomar.consolelogger
 
+import com.intellij.openapi.application.ApplicationManager
 import com.github.bgomar.bgconsolelogger.tools.ConsoleLoggerSettings
 import com.intellij.ide.plugins.PluginManagerCore
-import com.intellij.notification.Notification
-import com.intellij.notification.NotificationGroupManager
 import com.intellij.notification.NotificationType
 import com.intellij.openapi.extensions.PluginId
 import com.intellij.openapi.project.Project
 import com.intellij.openapi.startup.ProjectActivity
-import com.intellij.openapi.ui.Messages
 
 class UpdateNotify : ProjectActivity {
 
@@ -16,50 +14,34 @@ class UpdateNotify : ProjectActivity {
 
     override suspend fun execute(project: Project) {
         val settings = ConsoleLoggerSettings.getInstance()
-        if (settings.version == "Unknown") {
+        if (plugin.version == "Unknown" || plugin.version != settings.version) {
             settings.version = plugin.version
             showUpdate(project)
-        } else if (plugin.version != settings.version) {
+        } else if (plugin.version == settings.version) {
             settings.version = plugin.version
-            showUpdate(project)
         }
     }
 
     private fun showUpdate(project: Project) {
-        Messages.showInfoMessage("""
+        ApplicationManager.getApplication().invokeLater {
+            val notification = createNotification(
+                "ConsoleLogger plugin updated to version ${plugin.version}",
+                """
             🐛 Bugfix for Saving 😁👌<br><br>
             Sorry it took so long, it was quite a hard task.<br>
+            - ✅Added ChapterTool, to create chapters in your code <br>
+            - Double-click on the chapter to jump to the line <br>
+            - ✅Function Obfuscator shows in ToolWindow <br>
+            - Fixed the issue with the logline numbers <br>
             🤏 Small changes to defaults: <br>
-            - Changed input field sizes<br>
-            - Added second logger preset option<br>
-            - Fixed Issue with Code Insertion Position and <br>
-              Log Line Number Updates in Plugin<br>
-            - Added the Re-check line numbers function<br>
-            """, "ConsoleLogger plugin updated to version ${plugin.version}")
-
-        val notification = createNotification(
-            "ConsoleLogger plugin updated to version ${plugin.version}",
-            """
-            🐛 Bugfix for Saving 😁👌<br><br>
-            Sorry it took so long, it was quite a hard task.<br>
-            🤏 Small changes to defaults: <br>
-            - Changed input field sizes<br>
-            - Added second logger preset option<br>
-            - Fixed Issue with Code Insertion Position and <br>
-              Log Line Number Updates in Plugin<br>
-            - Added the Re-check line numbers function<br>
+            - Changed input field sizes <br>
+            - Fixed Issue with Code Insertion Position  <br>
+            - Added the Re-check line numbers function <br>
+            - Fixed issue with the removal of ConsoleLogger logs <br>
             """,
-            NotificationType.INFORMATION
-        )
-        showFullNotification(project, notification)
-    }
-
-    private fun createNotification(title: String, content: String, type: NotificationType): Notification {
-        val notificationGroup = NotificationGroupManager.getInstance().getNotificationGroup("com.github.bgomar.consolelogger.notification")
-        return notificationGroup.createNotification(title, content, type)
-    }
-
-    private fun showFullNotification(project: Project, notification: Notification) {
-        notification.notify(project)
+                NotificationType.INFORMATION
+            )
+            showFullNotification(project, notification)
+        }
     }
 }